added MyEvents tab; reorganized file structure
