+++ client/src/App.jsx
@@ -4,9 +4,7 @@ import ContextDemo from './components/ContextDemo';
 import axios from 'axios';
 import { saveAs } from 'file-saver';
 
-import Dashboard from './pages/Dashboard';
-import Login from './pages/Login';
-import Signup from './pages/Signup';
+import Dashboard from './pages/dashboard/Dashboard';
 
 import './App.css';
