+++ client/src/components/dashboardTabs/myEquip/myEquip.css
