+++ client/src/components/dashboardTabs/myEvents/eventDetails/EventDetails.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const eventDetails = () => {
+  return (
+    <div>
+      <h1>Event Details page</h1>
+    </div>
+  );
+};
+
+export default eventDetails;
