+++ client/src/components/dashboardTabs/myEvents/myEventsTab/MyEvents.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './myEvents.css';
+
+const MyEvents = () => {
+  return (
+    <div>
+      <h1>My Events Tab</h1>
+    </div>
+  );
+};
+
+export default MyEvents;
